radiusdtls-bis-09 - Detecting Live Servers - absence of reply
Clarify that the updated paragraph covers both the correct and incorrect cases.
That is, when the client draws a conclusion that the server is unavailable, the
conclusion may be correct or incorrect and the pargraph covers the both cases,
not just the problems with the incorrect conclusion.

I've so far understood the paragraph to only address the problems that arise
when the conclusion is incorrect. Janfred's comment clarified this is not the
case and I hope the updated text is also clearer about this.