radiusdtls-bis-09 - mostly DTLS related suggestions - (D)TLS data is opaque
Section named 'Default ports and shared secrets' has two paragraphs which
discuss non-(D)TLS data received by (D)TLS ports. This is better handled in
section '4. Packet / Connection Handling' which talks more about configuring
TLS server and clients to use (D)TLS, handshake, etc.

From a point of view of a RADIUS server implementator, removed text requires,
for example, discarding packets that are 'not a new DTLS negotiation'. For a
RADIUS server application this would require understanding and tracking DTLS to
a level that is required to understand if a packet is part of new negotiation
or not.

Text that is added by this commit says data received by a (D)TLS port is opaque
for a RADIUS client or server application. The opaque data is passsed to (D)TLS
implementation without further processing.